requirements.txt
asgiref==3.8.1
astroid==3.2.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.3.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Deprecated==1.2.14
dill==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
flake8==7.1.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
msgpack==1.0.8
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==1.7.1
pylint==3.2.5
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
requests==2.32.3
service-identity==24.1.0
setuptools==70.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stripe==10.1.0
tomli==2.0.1
tomlkit==0.12.5
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
wget==3.2
wrapt==1.16.0
zope.interface==6.4.post2
